rpmsg: allow RPMsg to be used indepdent to virtio
Add rpmsg_device struct to allow RPMsg to be used indepdent to
virtio.
